refactor(Player): 카드 컬렉션 포장
